lun_manager: Allow "create" on existing LUNs
When mode is set to "create" and the LUN already exists lun_manager
will verify that the existing LUN has all the required initiators.

If there are any initators that are missing, it'll add them.
If there are any "unknown" initators, lun_manager will refuse
to continue, with the assumption being that the LUN might belong
to another cluster, and we don't want to risk two clusters
accessing the same LUN.